frontend/pyproject.toml
[project]
name = "poc-apl-frontend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chainlit",
    "aiohttp",
    "python-dotenv",
    "aiosqlite",
    "sqlalchemy",
]

[tool.uv]
package = false
